requirements.txt
numpy==2.2.1
opencv-python==4.11.0.86
pyserial==3.5
scipy==1.15.1
